Fix navbar issue of not working
